+++ NEWS.md
@@ -4,6 +4,7 @@
 ## Minor changes and bug fixes
 - In `hierarchical_pw_clustering`, redefined the distance measure as `1 - kappa statistic`
 - Fixed minor issue in `cluster_graph_vis` (during the calculations for additional node colors)
+- Removed title from graph visualization of hierarchical clustering in `cluster_graph_vis`
 
 # pathfindR 1.3.0